permite renomear aldeias

+++ scripts/build-wasm.ps1
@@ -1,16 +1,30 @@
 # Copyright (C) Call of Nil contributors
 # SPDX-License-Identifier: AGPL-3.0-only
 
+param(
+  [switch]$Release
+)
+
 $ErrorActionPreference = 'Stop'
 $PSNativeCommandUseErrorActionPreference = $true
 
 function Build {
   param ([string] $Crate)
   
-  Get-ChildItem "crates/$Crate/pkg" -Recurse | Remove-Item
+  if (Test-Path -Path "crates/$Crate/pkg" ) {
+    Get-ChildItem "crates/$Crate/pkg" -Recurse | Remove-Item
+  }
 
-  wasm-pack build "crates/$Crate" --release --reference-types
+  $BuildCmd = "wasm-pack build crates/$Crate --reference-types"
+
+  if ($Release) {
+    $BuildCmd += ' --release'
+  }
+  else {
+    $BuildCmd += ' --dev --no-opt'
+  }
+
+  Invoke-Expression $BuildCmd
 }
-
 
 Build -Crate 'nil-continent'
